fix some architecture in models

+++ src/main/java/back/SportApp/Exercise/ExerciseServiceImpl.java
@@ -1,8 +1,11 @@
 package back.SportApp.Exercise;
 
+import back.SportApp.Training.Training;
+import back.SportApp.Training.TrainingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,9 +15,15 @@ public class ExerciseServiceImpl implements ExerciseService {
     @Autowired
     private ExerciseRepository exerciseRepository;
 
+    private TrainingRepository trainingRepository;
+
     @Override
     public Exercise create(Exercise exercise) {
-        return exerciseRepository.save(exercise);
+        if (!exerciseRepository.existsById(exercise.getId())) {
+            return exerciseRepository.save(exercise);
+        } else {
+            throw new RuntimeException("Training not found");
+        }
     }
 
     @Override
@@ -23,10 +32,9 @@ public class ExerciseServiceImpl implements ExerciseService {
     }
 
     @Override
-    public Exercise findById(Long id) {
+    public Exercise findById(Integer id) {
         Optional<Exercise> exercise = exerciseRepository.findById(id);
         if (exercise.isPresent()) {
-
             return exercise.get();
         } else {
             throw new RuntimeException(("Exercise not found with id "+ id)) ;
@@ -35,14 +43,14 @@ public class ExerciseServiceImpl implements ExerciseService {
 
     @Override
     public void update(Exercise exercise) {
-        final Long id = exercise.getId();
+        final Integer id = exercise.getId();
         Optional<Exercise> existingExercise = exerciseRepository.findById(id);
         if (existingExercise.isPresent()) {
             Exercise updateExercise = existingExercise.get();
             updateExercise.setName(exercise.getName());
             updateExercise.setRep(exercise.getRep());
             updateExercise.setSet(exercise.getSet());
-            updateExercise.setTraining(exercise.getTraining());
+            updateExercise.setTrainings(exercise.getTrainings());
             exerciseRepository.save(updateExercise);
         } else {
             throw new RuntimeException("Wrong or inexistant ID" + id);
@@ -50,8 +58,12 @@ public class ExerciseServiceImpl implements ExerciseService {
     }
 
     @Override
-    public void deleteById(Long id) {
-        exerciseRepository.deleteById(id);
+    public void deleteById(Integer id) {
+        final Exercise exercise = this.findById(id);
+        if(exercise != null) {
+            exerciseRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Wrong or inexistant ID" + id);
+        }
     }
-
 }
